chore: update to new multiformats
I had to skip a couple of tests as they use the ipfs api directly which
hasn't been migrated yet so it doesn't understand the new CID class.

They can be unskipped once js-ipfs ships with the new multiformats module,
or the circular dependency between this module and js-ipfs is broken.

BREAKING CHANGE: pulls in new multiformats modules